Move Vedeu::Model::Collection and Vedeu::Model;
Move Vedeu::Model::Collection to Vedeu::Collection.
Move Vedeu::Collection to repositories/.
Move Vedeu::Model to repositories/.
Move vedeu/models/view/collections to vedeu/repositories/collections.
Update require paths.
