+++ lib/vedeu/cursor/cursor.rb
@@ -1,5 +1,4 @@
 require 'vedeu/cursor/all'
-require 'vedeu/models/model'
 require 'vedeu/geometry/position'
 require 'vedeu/support/visible'
