+++ lib/vedeu/models/all.rb
@@ -1,12 +1,10 @@
-require 'vedeu/models/collection'
-
+require 'vedeu/repositories/all'
 require 'vedeu/models/view/all'
 
 require 'vedeu/models/focus'
 require 'vedeu/models/geometry'
 require 'vedeu/models/group'
 require 'vedeu/models/menu'
-require 'vedeu/models/model'
 
 module Vedeu
 end # Vedeu
