+++ lib/vedeu/repositories/collections.rb
@@ -1,35 +1,35 @@
-require 'vedeu/models/collection'
+require 'vedeu/repositories/collection'
 
 module Vedeu
 
   # A collection of {Vedeu::Char} instances.
   #
-  class Chars < Vedeu::Model::Collection
+  class Chars < Vedeu::Collection
   end # Chars
 
   # A collection of {Vedeu::Event} instances.
   #
-  class Events < Vedeu::Model::Collection
+  class Events < Vedeu::Collection
   end # Events
 
   # A collection of {Vedeu::Interface} instances.
   #
-  class Interfaces < Vedeu::Model::Collection
+  class Interfaces < Vedeu::Collection
   end # Interfaces
 
   # A collection of {Vedeu::Key} instances.
   #
-  class Keys < Vedeu::Model::Collection
+  class Keys < Vedeu::Collection
   end # Keys
 
   # A collection of {Vedeu::Line} instances.
   #
-  class Lines < Vedeu::Model::Collection
+  class Lines < Vedeu::Collection
   end # Lines
 
   # A collection of {Vedeu::Stream} instances.
   #
-  class Streams < Vedeu::Model::Collection
+  class Streams < Vedeu::Collection
   end # Streams
 
 end # Vedeu
