+++ lib/vedeu/repositories/model.rb
