+++ test/lib/vedeu/repositories/model_test.rb
